# openapi.yaml
openapi: 3.0.3
info:
  title: HLS Streaming Server API
  description: |
    Complete REST API for managing HLS streaming channels, media libraries, and dynamic playlists.

    ## Features
    - Multi-channel HLS streaming with FFmpeg
    - Dynamic playlist generation via schedule blocks
    - Jellyfin-style library management
    - Media bucket organization
    - EPG/XMLTV generation
    - IPTV M3U playlist export
    - Session-based authentication
    - Virtual time tracking for 24/7 simulation

    ## Authentication
    The API supports two authentication methods:
    1. **API Key** (Legacy): Pass via `X-API-Key` header or `apiKey` query parameter
    2. **Session Token** (Modern): Pass via `Authorization: Bearer <token>` header or `sessionToken` cookie

    Most endpoints require authentication except for streaming and public EPG endpoints.

  version: 1.0.0
  contact:
    name: HLS Streaming Server
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-server.com
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionToken
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
      required:
        - success
...
    ScheduleBlock:
      type: object
      properties:
        id:
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        name:
          type: string
        dayOfWeek:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          description: "0=Sunday, 6=Saturday"
        startTime:
          type: string
          pattern: '^\d{2}:\d{2}:\d{2}$'
          example: "20:00:00"
        endTime:
          type: string
          pattern: '^\d{2}:\d{2}:\d{2}$'
          example: "23:00:00"
        bucketId:
          type: string
          format: uuid
        playbackMode:
          type: string
          enum: [sequential, random, shuffle]
        priority:
          type: integer
        enabled:
          type: boolean
    CreateScheduleBlockRequest:
      type: object
      properties:
        name:
          type: string
          example: "Prime Time"
        dayOfWeek:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          nullable: true
        startTime:
          type: string
          pattern: '^\d{2}:\d{2}:\d{2}$'
          example: "20:00:00"
        endTime:
          type: string
          pattern: '^\d{2}:\d{2}:\d{2}$'
          example: "23:00:00"
        bucketId:
          type: string
          format: uuid
        playbackMode:
          type: string
          enum: [sequential, random, shuffle]
        priority:
          type: integer
        enabled:
          type: boolean
          default: true
      required:
        - name
        - startTime
        - endTime
        - bucketId
        - playbackMode
        - priority
...
paths:
...
  /api/schedules/channels/{channelId}/blocks:
    get:
      tags: [Schedules]
      summary: List schedule blocks
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Schedule blocks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ScheduleBlock'

    post:
      tags: [Schedules]
      summary: Create schedule block
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleBlockRequest'
      responses:
        '200':
          description: Schedule block created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/schedules/channels/{channelId}/blocks/{blockId}:
    get:
      tags: [Schedules]
      summary: Get schedule block
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: blockId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Schedule block
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    put:
      tags: [Schedules]
      summary: Update schedule block
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: blockId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleBlockRequest'
      responses:
        '200':
          description: Schedule block updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags: [Schedules]
      summary: Delete schedule block
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: blockId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Schedule block deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'